eID加密机添加，eID表结构更新

+++ eid-dal/src/main/java/com/eid/dal/dao/CompanyAppDao.java
@@ -16,7 +16,7 @@ public interface CompanyAppDao extends CrudRepository<CompanyAppEntity, Long> {
 
     CompanyAppEntity findByApId(String apId);
 
-    CompanyAppEntity findById(int id);
+    CompanyAppEntity findById(Long id);
 
     @Modifying
     @Transactional
